slight fix on stopping during recess

+++ src/PomodoroTimer/PomodoroTimer.js
@@ -9,6 +9,7 @@ let states={
     off: {"bg_color":"white"},
     work: {"bg_color":"#86db8d"},
     recess: {"bg_color":"#7b9fbd"},
+    pause: {"bg_color":"#FFFF00"},
 }
 
 export default function PomodoroTimer(){
@@ -17,21 +18,29 @@ export default function PomodoroTimer(){
     let [ interval_time, setInterval_time ] = useState(0)
     let [ change, setChange ] = useState("sup")
     let [ state, setState ] = useState("off")
+    let [ from, setFrom ] = useState("off")
     let [ intervals_completed, setIntervals_completed ] = useState(0)
 
     useEffect(() => {
+        //window.addEventListener("keydown", (e) => {
+            //console.log('test')
+        //})
         if(!worker){
             setWorker(new Worker(new URL('./worker.js',import.meta.url)))
+            console.log('test')
         }
         //switches between work and recess
         if(change == 1){
+            console.log('test2')
             setChange(0)
             audio.play()
+            console.log(state)
+            console.log(from)
 
             //if == recess cause we got from recess -> work and work -> recess
             let time_to_set = state == "recess" ? 
                 interval_time : 
-                interval_time / TIME_TO_RECESS; 
+                interval_time / 2; 
 
             if (state == 'work') {
                 setState('recess')
@@ -47,19 +56,22 @@ export default function PomodoroTimer(){
     },[worker, change])
 
     function stopWorker(){
+        if(state=="off") return
         worker.terminate()
+        setFrom(state)
         setState("off")
         setWorker(null)
     }
 
 	function startWorker(){
         if(state == 'work') return
-        setState("work")
+        from == "recess" ? setState("recess") : setState("work")
 		worker.postMessage({
             time:time==0?interval_time:time,
 		})
 		worker.onmessage = (e) =>{
             setTime(e.data)
+            console.log(e.data)
             if(e.data == 0) setChange(1)
 		}
 	}
